backend/packages/hsm_api/pyproject.toml
[project]
name = "hsm-api"
version = "0.1.0"
description = "Habitat Suitability Model Visualiser API"
requires-python = ">=3.11"
dependencies = [
    "hsm-core",
    "firebase-admin>=6.9.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-tasks>=2.14.0",
    "httpx>=0.28.0",
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.11.5",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
    "rasterio>=1.4.4",
    "numpy>=1.26.0",
    "pandas[parquet]>=2.0.0",
    "pyarrow>=14.0.0",
    "scikit-learn>=1.5.2",
    "shap>=0.44.0",
    "elapid>=1.0.2",
]

[tool.uv.sources]
hsm-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend_api"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
]
